False Position Method
Theory:
The False Position Method (Regula Falsi) is a numerical method used to approximate the 
root of a function. 
It is similar to the Bisection Method, but instead of taking the midpoint, it finds the 
intersection of the secant line with the x-axis
Formula: 
Xr=xl⋅f(xu)−xu⋅f(xl)/f(xu)−f(xl) 
Steps: 
1. Choose initial guesses xl and xu such that f(xl)⋅f(xu)<0.
2. Compute root approximation using the formula above.
3. If f(xr)= or error < tolerance, stop.
4. Otherwise, replace xlor xu depending on the sign of f(xr)..
5. Repeat until desired accuracy is achieved.
The false position method often converges faster than the bisection method.